posmap final

+++ src/runtime/local/io/utils.h
@@ -34,10 +34,10 @@ ValueTypeCode inferValueType(const std::string &value);
 FileMetaData generateFileMetaData(const std::string &filename, char delim, size_t sampleRows, bool isMatrix = false);
 
 // Function to create and save the positional map
-void writePositionalMap(const char *filename, const std::vector<std::pair<std::streampos, std::vector<uint32_t>>> &posMap);
+void writePositionalMap(const char *filename, const std::vector<std::pair<std::streampos, std::vector<uint16_t>>> &posMap);
 
 // Function to read the positional map
-std::vector<std::pair<std::streampos, std::vector<uint32_t>>> readPositionalMap(const char *filename);
+std::vector<std::pair<std::streampos, std::vector<uint16_t>>> readPositionalMap(const char *filename);
 
 // Conversion of std::string.
